+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v6.28](https://github.com/realityforge/replicant/tree/v6.28) (2019-02-07)
+[Full Changelog](https://github.com/realityforge/replicant/compare/v6.27...v6.28)
 
 * Update the `org.realityforge.arez` dependencies to version `0.128`.
 * Update the `org.realityforge.react4j` dependencies to version `0.117`.